backend/requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.13
# by the following command:
#
#    pip-compile --output-file=requirements.txt requirements.in
#
blinker==1.9.0
    # via flask
brotli==1.2.0
    # via fonttools
cachelib==0.13.0
    # via flask-session
cffi==2.0.0
    # via weasyprint
chardet==5.2.0
    # via reportlab
click==8.3.0
    # via flask
colorama==0.4.6
    # via
    #   click
    #   pytest
cssselect2==0.8.0
    # via weasyprint
et-xmlfile==2.0.0
    # via openpyxl
flask==3.1.1
    # via
    #   -r requirements.in
    #   flask-cors
    #   flask-jwt-extended
    #   flask-session
    #   flask-sqlalchemy
    #   pytest-flask
flask-cors==6.0.0
    # via -r requirements.in
flask-jwt-extended==4.7.1
    # via -r requirements.in
flask-session==0.8.0
    # via -r requirements.in
flask-sqlalchemy==3.1.1
    # via -r requirements.in
fonttools[woff]==4.60.1
    # via weasyprint
gunicorn==23.0.0
    # via -r requirements.in
iniconfig==2.1.0
    # via pytest
itsdangerous==2.2.0
    # via flask
jinja2==3.1.6
    # via flask
markupsafe==3.0.3
    # via
    #   flask
    #   jinja2
    #   werkzeug
msgspec==0.19.0
    # via flask-session
nodeenv==1.9.1
    # via pyright
openpyxl==3.1.5
    # via -r requirements.in
packaging==25.0
    # via
    #   gunicorn
    #   pytest
pillow==12.0.0
    # via
    #   reportlab
    #   weasyprint
pluggy==1.6.0
    # via pytest
psutil==6.1.1
    # via -r requirements.in
psycopg2-binary==2.9.10
    # via -r requirements.in
pycparser==2.23
    # via cffi
pydyf==0.11.0
    # via weasyprint
pyjwt==2.10.1
    # via
    #   -r requirements.in
    #   flask-jwt-extended
pyphen==0.17.2
    # via weasyprint
pyright==1.1.390
    # via -r requirements.in
pytest==8.3.5
    # via
    #   -r requirements.in
    #   pytest-flask
pytest-flask==1.3.0
    # via -r requirements.in
python-barcode==0.15.1
    # via -r requirements.in
python-dotenv==1.0.1
    # via -r requirements.in
python-json-logger==3.2.1
    # via -r requirements.in
reportlab==4.2.5
    # via -r requirements.in
ruff==0.14.2
    # via -r requirements.in
segno==1.6.1
    # via -r requirements.in
sqlalchemy==2.0.36
    # via
    #   -r requirements.in
    #   flask-sqlalchemy
tinycss2==1.4.0
    # via
    #   cssselect2
    #   weasyprint
tinyhtml5==2.0.0
    # via weasyprint
typing-extensions==4.15.0
    # via
    #   pyright
    #   sqlalchemy
weasyprint==63.1
    # via -r requirements.in
webencodings==0.5.1
    # via
    #   cssselect2
    #   tinycss2
    #   tinyhtml5
werkzeug==3.1.3
    # via
    #   -r requirements.in
    #   flask
    #   flask-cors
    #   flask-jwt-extended
    #   pytest-flask
zopfli==0.2.3.post1
    # via fonttools
